# docs/swagger.yaml
basePath: /swagger/datahub-service
definitions:
  controllers.APIResponse:
    properties:
      data: {}
      msg:
        example: 操作成功
        type: string
      status:
        example: 0
        type: integer
    type: object
...
  controllers.BatchDeleteRequest:
    properties:
      task_ids:
        items:
          type: string
        type: array
    required:
    - task_ids
    type: object
...
  controllers.SyncTaskCreateRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      created_by:
        example: admin
        type: string
      data_source_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      interface_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      library_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      library_type:
        example: basic_library
        type: string
      task_type:
        example: full_sync
        type: string
    required:
    - data_source_id
    - library_id
    - library_type
    - task_type
    type: object
  controllers.SyncTaskUpdateRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      updated_by:
        example: admin
        type: string
    type: object
...
  models.BasicLibrary:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      data_sources:
        items:
          $ref: '#/definitions/models.DataSource'
        type: array
      description:
        example: 存储用户基础信息的数据库
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      interfaces:
        description: 关联关系
        items:
          $ref: '#/definitions/models.DataInterface'
        type: array
      name_en:
        example: user_basic_library
        type: string
      name_zh:
        example: 用户数据基础库
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.DataInterface:
    properties:
      basic_library:
        allOf:
        - $ref: '#/definitions/models.BasicLibrary'
        description: 关联关系
      clean_rules:
        items:
          $ref: '#/definitions/models.CleansingRule'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      data_source:
        $ref: '#/definitions/models.DataSource'
      data_source_id:
        type: string
      description:
        type: string
      id:
        type: string
      interface_config:
        $ref: '#/definitions/models.JSONB'
      is_table_created:
        type: boolean
      library_id:
        type: string
      name_en:
        type: string
      name_zh:
        type: string
      parse_config:
        $ref: '#/definitions/models.JSONB'
      status:
        type: string
      table_fields_config:
        $ref: '#/definitions/models.JSONB'
      type:
        description: realtime, batch
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.SyncProgress:
    properties:
      current_phase:
        type: string
      error_count:
        type: integer
      estimated_time:
        type: string
      processed_rows:
        type: integer
      progress_percent:
        type: integer
      speed:
        description: 每秒处理行数
        type: integer
      total_rows:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.SyncTask:
    properties:
      basic_library:
        allOf:
        - $ref: '#/definitions/models.BasicLibrary'
        description: 动态关联 - 这些字段不存储在数据库中，在运行时根据LibraryType动态加载
      config:
        additionalProperties: true
        description: 同步配置
        type: object
      created_at:
        type: string
      created_by:
        example: system
        type: string
      data_interface:
        $ref: '#/definitions/models.DataInterface'
      data_source:
        $ref: '#/definitions/models.DataSource'
      data_source_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      end_time:
        type: string
      error_count:
        example: 0
        type: integer
      error_message:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      interface_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      library_id:
        description: 基础库ID或主题库ID
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      library_type:
        description: basic_library, thematic_library
        example: basic_library
        type: string
      processed_rows:
        example: 0
        type: integer
      progress:
        description: 进度百分比 0-100
        example: 0
        type: integer
      result:
        additionalProperties: true
        description: 同步结果
        type: object
      start_time:
        type: string
      status:
        description: pending, running, success, failed, cancelled
        example: pending
        type: string
      task_type:
        description: full_sync, incremental_sync, realtime_sync
        example: full_sync
        type: string
      thematic_library:
        $ref: '#/definitions/models.ThematicLibrary'
      total_rows:
        example: 0
        type: integer
      updated_at:
        type: string
    type: object
...
  service.BatchDeleteResponse:
    properties:
      deleted_count:
        type: integer
      errors:
        items:
          type: string
        type: array
      failed_ids:
        items:
          type: string
        type: array
    type: object
  service.PaginationInfo:
    properties:
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  service.SyncTaskListResponse:
    properties:
      pagination:
        $ref: '#/definitions/service.PaginationInfo'
      tasks:
        items:
          $ref: '#/definitions/models.SyncTask'
        type: array
    type: object
  service.SyncTaskStatistics:
    properties:
      cancelled_tasks:
        type: integer
      failed_tasks:
        type: integer
      pending_tasks:
        type: integer
      running_tasks:
        type: integer
      success_rate:
        type: number
      success_tasks:
        type: integer
      total_tasks:
        type: integer
    type: object
  service.SyncTaskStatusResponse:
    properties:
      error:
        type: string
      processor:
        type: string
      progress:
        $ref: '#/definitions/models.SyncProgress'
      result:
        $ref: '#/definitions/models.SyncResult'
      start_time:
        type: string
      status:
        type: string
      task:
        $ref: '#/definitions/models.SyncTask'
    type: object
info:
  contact: {}
  description: 智慧园区数据底座后台服务，提供数据采集、处理、存储、治理和共享功能
  title: 数据底座服务 API
  version: "1.0"
paths:
...
  /sync/tasks:
    get:
      consumes:
      - application/json
      description: |-
        分页获取同步任务列表，支持多种过滤条件

        **查询参数说明:**
        - page: 页码，默认1
        - size: 每页大小，默认10，最大100
        - library_type: 库类型过滤
        - library_id: 库ID过滤
        - data_source_id: 数据源ID过滤
        - status: 任务状态过滤
        - task_type: 任务类型过滤
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页大小
        in: query
        name: size
        type: integer
      - description: 库类型
        in: query
        name: library_type
        type: string
      - description: 库ID
        in: query
        name: library_id
        type: string
      - description: 数据源ID
        in: query
        name: data_source_id
        type: string
      - description: 任务状态
        in: query
        name: status
        type: string
      - description: 任务类型
        in: query
        name: task_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.SyncTaskListResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取同步任务列表
      tags:
      - 同步任务管理
    post:
      consumes:
      - application/json
      description: |-
        创建新的数据同步任务，支持基础库和主题库

        **支持的库类型:**
        - basic_library: 基础库
        - thematic_library: 主题库

        **支持的任务类型:**
        - full_sync: 全量同步
        - incremental_sync: 增量同步
        - realtime_sync: 实时同步

        **任务状态流转:**
        pending → running → success/failed/cancelled
      parameters:
      - description: 同步任务创建信息
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/controllers.SyncTaskCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SyncTask'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建同步任务
      tags:
      - 同步任务管理
  /sync/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的同步任务，只能删除已完成、失败或已取消的任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "409":
          description: 任务状态不允许删除
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除同步任务
      tags:
      - 同步任务管理
    get:
      consumes:
      - application/json
      description: 根据任务ID获取同步任务的详细信息，包括关联的库信息、数据源信息等
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SyncTask'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取同步任务详情
      tags:
      - 同步任务管理
    put:
      consumes:
      - application/json
      description: 更新同步任务的配置信息，只能更新处于 pending 状态的任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/controllers.SyncTaskUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SyncTask'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "409":
          description: 任务状态不允许更新
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 更新同步任务
      tags:
      - 同步任务管理
  /sync/tasks/{id}/cancel:
    post:
      consumes:
      - application/json
      description: 取消指定的同步任务，可以取消待执行或正在执行的任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 取消成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "409":
          description: 任务状态不允许取消
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 取消同步任务
      tags:
      - 同步任务管理
  /sync/tasks/{id}/retry:
    post:
      consumes:
      - application/json
      description: 重试失败的同步任务，会创建一个新的任务实例
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 重试成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SyncTask'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "409":
          description: 任务状态不允许重试
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 重试同步任务
      tags:
      - 同步任务管理
  /sync/tasks/{id}/start:
    post:
      consumes:
      - application/json
      description: 启动指定的同步任务，将任务提交给同步引擎执行
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 启动成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "409":
          description: 任务状态不允许启动
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 启动同步任务
      tags:
      - 同步任务管理
  /sync/tasks/{id}/status:
    get:
      consumes:
      - application/json
      description: 获取同步任务的实时执行状态和进度信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.SyncTaskStatusResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取同步任务状态
      tags:
      - 同步任务管理
  /sync/tasks/{id}/stop:
    post:
      consumes:
      - application/json
      description: 停止正在执行的同步任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 停止成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "409":
          description: 任务状态不允许停止
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 停止同步任务
      tags:
      - 同步任务管理
  /sync/tasks/batch-delete:
    post:
      consumes:
      - application/json
      description: 批量删除多个同步任务，只能删除已完成、失败或已取消的任务
      parameters:
      - description: 批量删除请求
        in: body
        name: tasks
        required: true
        schema:
          $ref: '#/definitions/controllers.BatchDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.BatchDeleteResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 批量删除同步任务
      tags:
      - 同步任务管理
  /sync/tasks/statistics:
    get:
      consumes:
      - application/json
      description: 获取同步任务的统计数据，包括各状态任务数量、成功率等
      parameters:
      - description: 库类型过滤
        in: query
        name: library_type
        type: string
      - description: 库ID过滤
        in: query
        name: library_id
        type: string
      - description: 数据源ID过滤
        in: query
        name: data_source_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.SyncTaskStatistics'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取同步任务统计信息
      tags:
      - 同步任务管理
...
swagger: "2.0"